warn instead of error if copy file fails, then assert success at end

+++ mdocean/plots/move_results_to_pubs.m
@@ -5,26 +5,34 @@ function move_results_to_pubs()
     RE = 'pubs/renewable-energy-mdo';
     AOR = 'pubs/applied-ocean-research-model';
 
+    success = true;
+
     for i = 1:length(figs_in_RE)
-        copy_result(figs_in_RE{i}, '.pdf', RE);
+        success = success && copy_result(figs_in_RE{i}, '.pdf', RE);
     end
     for i = 1:length(figs_in_AOR)
-        copy_result(figs_in_AOR{i}, '.pdf', AOR);
+        success = success && copy_result(figs_in_AOR{i}, '.pdf', AOR);
     end
     for i = 1:length(tabs_in_RE)
-        copy_result(tabs_in_RE{i}, '.tex', RE);
+        success = success && copy_result(tabs_in_RE{i}, '.tex', RE);
     end
     for i = 1:length(tabs_in_AOR)
-        copy_result(tabs_in_AOR{i}, '.tex', AOR);
+        success = success && copy_result(tabs_in_AOR{i}, '.tex', AOR);
     end
-
+    assert(success, 'Some files were not found and could not be copied.')
 end
 
-function copy_result(class_dot_description, extension, dest_folder)
+function success = copy_result(class_dot_description, extension, dest_folder)
     s = split(class_dot_description, '.');
     class_name = s{1};
     fig_tab_desc = s{2};
     src = fullfile('results', class_name, [fig_tab_desc, extension]);
     dest = fullfile(dest_folder, [fig_tab_desc, extension]);
-    copyfile(src, dest);
+    if exist(src,"file")
+        copyfile(src, dest);
+        success = true;
+    else
+        warning(['File not found: ' src])
+        success = false;
+    end
 end
